fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.1.0
langchain==0.1.12
langchain-google-genai==0.0.11
google-generativeai>=0.4.1,<0.5.0
beautifulsoup4==4.12.3
chromadb==0.4.24
sentence-transformers==2.5.1
firebase-admin==6.5.0
requests==2.31.0